// WebApplicationPlateforme/ClientApp/src/app/shared/Models/MediaCenter/RecordingArchive/recording-archive.model.ts
export class RecordingArchive {

  id: number;
  type: string;
  but: string;
  localisation: string;
  dateTime: string;
  budget: string;
  detail: string;

  dirid: string;
  dirnom: string;
  diretat: string;
  dirdate: string;
  adminid: string;
  adminom: string;
  admietat: string;
  admidate: string;
  etabnom: string;
  etabid: string;
  etabetat: string;
  etabdate: string;
  attribut1: string;
  attribut2: string;
  attribut3: string;
  attribut4: string;
  attribut5: string;
  attribut6: string;
  dateenreg: string;
  userNameCreator: string;
  idUserCreator: string;

}